fix: sys.database_query_store_options storage size columns are bigint
current_storage_size_mb/max_storage_size_mb were read via GetDecimal, but
they're bigint, not decimal - caused an InvalidCastException on every
--self-test run, in both RunQueryStoreStatus and RunQueryStoreTopQueries
(the latter calls the former internally to check state before pulling top
queries, so one bad read failed both). GetInt64 converts implicitly to the
decimal model fields, so no model change needed.

Found by running --self-test against a real instance - the two Query Store
checks were the only failures out of 140 steps, down from 4/126 last run.

+++ HealthCheckAnalyzer.cs
@@ -747,8 +747,8 @@ public sealed class HealthCheckAnalyzer
             ActualStateDesc = r.GetString(0),
             DesiredStateDesc = r.GetString(1),
             ReadonlyReason = r.GetInt32(2),
-            CurrentStorageSizeMb = r.GetDecimal(3),
-            MaxStorageSizeMb = r.GetDecimal(4),
+            CurrentStorageSizeMb = r.GetInt64(3),
+            MaxStorageSizeMb = r.GetInt64(4),
             QueryCaptureModeDesc = r.GetString(5)
         };
     }
